api/requirements.txt
flask
flask-sqlalchemy
flask-bcrypt
python-dotenv